bugfix: Use proper conversion with physics simulation
The macOS compiler isn't as nice as the ubuntu one

+++ examples/mini/physics_simulation/floor.cpp
@@ -28,10 +28,10 @@ void Floor::update(f32 delta) {
 
 void Floor::render(LTEngine::Rendering::Renderer *renderer) {
 	renderer->setZOrder(1);
-	renderer->drawRect({0, 0, (u32)m_size.x, (u32)m_size.y}, LTEngine::Rendering::ColorA(30, 30, 30, 255));
+	renderer->drawRect({0, 0, (f32)m_size.x, (f32)m_size.y}, LTEngine::Rendering::ColorA(30, 30, 30, 255));
 }
 
 
 std::unique_ptr<LTEngine::Object::ObjectStructure::Object> Floor::clone() const {
-	return std::make_unique<Floor>(m_physicsEngine, (LTEngine::Shapes::Rect){position.x, position.y, (u32)m_size.x, (u32)m_size.y});
+	return std::make_unique<Floor>(m_physicsEngine, (LTEngine::Shapes::Rect){position.x, position.y, (f32)m_size.x, (f32)m_size.y});
 }
